// src/models/cards.ts
export interface Card{
    telefon: string;
    telefax: string;
    occupation: string;
    address1: string;
    address2: string;
    address3: string;
    company: string;
    website: string;
    email: string;
    description: string;
    color: string;
    display_name: string;
  }
  